requirements.txt
torch
torchvision
zarr
pandas
jupyter
tqdm
python-dotenv
wandb
matplotlib
cartopy
xarray
ipympl
lightning